requirements.txt
asgiref==3.8.0
Django==5.0.3
djangorestframework==3.15.0
lxml==5.1.0
pillow==10.2.0
python-docx==1.1.0
sqlparse==0.4.4
typing_extensions==4.10.0
